PRO-5389: fixed dupli code
